requirements.txt
dict.cc.py
